rename `internal/test_utils/compare/two_stage_l3_clos/security_zone.go` -> `internal/test_utils/compare/datacenter/security_zone.go`
