+++ internal/test_utils/compare/datacenter/security_zone.go
